Add Database.cs facade functionality
A new file, Database.cs, is created in the GameEditor/References directory. This file contains a static Database class that provides a facade around the LocalbaseDatabase. The class includes functions for adding/removing listeners, and setting/getting values from the database. This provides a higher-level, more convenient interface for interacting with the database.