Refactor database interaction in ProjectSceneBehaviour and Database classes
This commit makes significant changes to the way the `ProjectSceneBehaviour` and `Database` classes handle interactions with the database. It removes the `_projectsDatabaseReference` object from the former and employs the new methods in the latter class. All previous Firebase-related code snippets were replaced by more concise and readable methods. Also, the entire process for listening and responding to changes in data is now more intuitive and simpler.