Implement major refactor to database handling
Refactored how data is retrieved and handled from the database. Simplified the database operation by aggregating function calls into a Database class, eliminating multiple direct calls to the database. In the process, HttpCache was initialized in the LocalbaseFileMonitor and in the ProjectSingletonBehaviour class, the setting of the project GUID was revised. This refactoring aims to centralize and simplify database connections and operations.