Add serialization utilities and improve test objects in Unity
The commit adds several new serialization utilities such as MeshFilterJsonConverter and updates the GameObjectJsonConverter. It also introduces a UnityObjectReference class. Multiple test objects were adjusted or created within the DefaultAssetPack asset and corresponding serialized data was updated in the DefaultAssetPack.json. This will allow for more thorough testing of serialization and referencing components in Unity.